Updated XIA Handel and Xerxes support for MX 1.2.0
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@313 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/i_xia_xerxes.c
@@ -860,7 +860,7 @@ mxi_xia_xerxes_restore_config( MX_XIA_XERXES *xia_xerxes )
 					mca_record->record_type_struct;
 
 				xia_dxp_mca->old_preset_type
-					= (uint32_t) MX_ULONG_MAX;
+					= (unsigned long) MX_ULONG_MAX;
 
 				mca->new_data_available = TRUE;
 				xia_dxp_mca->new_statistics_available = TRUE;
@@ -1092,7 +1092,7 @@ mxi_xia_xerxes_is_busy( MX_MCA *mca,
 MX_EXPORT mx_status_type
 mxi_xia_xerxes_read_parameter( MX_MCA *mca,
 			char *parameter_name,
-			uint32_t *value_ptr,
+			unsigned long *value_ptr,
 			int debug_flag )
 {
 	static const char fname[] = "mxi_xia_xerxes_read_parameter()";
@@ -1147,7 +1147,7 @@ mxi_xia_xerxes_read_parameter( MX_MCA *mca,
 			mxi_xia_xerxes_strerror( xia_status ) );
 	}
 
-	*value_ptr = (uint32_t) short_value;
+	*value_ptr = (unsigned long) short_value;
 
 	if ( debug_flag ) {
 		MX_DEBUG(-2,("%s: value = %lu", fname, *value_ptr));
@@ -1159,7 +1159,7 @@ mxi_xia_xerxes_read_parameter( MX_MCA *mca,
 MX_EXPORT mx_status_type
 mxi_xia_xerxes_write_parameter( MX_MCA *mca,
 			char *parameter_name,
-			uint32_t value,
+			unsigned long value,
 			int debug_flag )
 {
 	static const char fname[] = "mxi_xia_xerxes_write_parameter()";
@@ -1238,10 +1238,11 @@ mxi_xia_xerxes_write_parameter( MX_MCA *mca,
 MX_EXPORT mx_status_type
 mxi_xia_xerxes_write_parameter_to_all_channels( MX_MCA *mca,
 			char *parameter_name,
-			uint32_t value,
+			unsigned long value,
 			int debug_flag )
 {
-	static const char fname[] = "mxi_xia_xerxes_write_parameter_to_all_channels()";
+	static const char fname[] =
+		"mxi_xia_xerxes_write_parameter_to_all_channels()";
 
 	MX_XIA_DXP_MCA *xia_dxp_mca;
 	MX_XIA_XERXES *xia_xerxes;
@@ -1631,9 +1632,9 @@ mxi_xia_xerxes_read_statistics( MX_MCA *mca,
 			xia_dxp_mca->output_count_rate, nevent ));
 	}
 
-	xia_dxp_mca->num_events = (uint32_t) nevent;
+	xia_dxp_mca->num_events = (unsigned long) nevent;
 
-	xia_dxp_mca->num_fast_peaks = (uint32_t) mx_round(
+	xia_dxp_mca->num_fast_peaks = (unsigned long) mx_round(
 				xia_dxp_mca->input_count_rate * mca->live_time);
 
 	xia_dxp_mca->num_underflows = 0;
@@ -1662,8 +1663,8 @@ mxi_xia_xerxes_read_statistics( MX_MCA *mca,
 	if ( mx_status.code != MXE_SUCCESS )
 		return mx_status;
 
-	xia_dxp_mca->num_underflows = (uint32_t) parameter_array[0]
-				+ 65536L * (uint32_t) parameter_array[1];
+	xia_dxp_mca->num_underflows = (unsigned long) parameter_array[0]
+				+ 65536L * (unsigned long) parameter_array[1];
 
 	mx_status = mxi_xia_xerxes_read_dspsymbol_array( xia_dxp_mca,
 					"OVERFLOWS", 2, parameter_array );
@@ -1671,8 +1672,8 @@ mxi_xia_xerxes_read_statistics( MX_MCA *mca,
 	if ( mx_status.code != MXE_SUCCESS )
 		return mx_status;
 
-	xia_dxp_mca->num_overflows = (uint32_t) parameter_array[0]
-				+ 65536L * (uint32_t) parameter_array[1];
+	xia_dxp_mca->num_overflows = (unsigned long) parameter_array[0]
+				+ 65536L * (unsigned long) parameter_array[1];
 #endif
 
 	return MX_SUCCESSFUL_RESULT;
